requirements.txt
PyJWT==2.6.0
tornado==6.2
